friday

+++ app/Http/Controllers/UserTypeController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\UserType;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Session;
 use Validator;
 use Input;
 use BF;
@@ -12,6 +14,50 @@ class UserTypeController extends Controller
 {
     public function index()
     {
+        $perm = Session::get('perm');
+        $cols = [
+            'id',
+            'name',
+            'deleted_at'
+        ] ;
+        $data = [];
+//        if (BF::getPermission('usertype.edit')) {
+//            $sql = UserType::withTrashed()->select($cols);
+//        } else {
+//            $sql = UserType::select($cols);
+//        }
+
+        $sql = UserType::select($cols);
+
+        // -- Order
+        foreach (json_decode(Input::get('order')) as $order) {
+            $sql->orderBy($order->column, $order->dir);
+        }
+
+        // -- Filter
+        foreach (json_decode(Input::get('columns')) as $col) {
+            $column = $col->data;
+            $val = $col->search;
+            if (in_array($column, $cols) && ( $val != '') ) {
+                $sql->where($column, 'LIKE', '%' . $val . '%');
+            }
+        }
+
+//        DB::enableQueryLog();
+//        $sql->get();
+//        $query = DB::getQueryLog();
+//        return $query ;
+//        exit();
+        try {
+            $count = $sql->count();
+            $data = $sql->skip(Input::get('start'))->take(Input::get('length'))->get();
+            $result = BF::dataTable($data, $count, $count, $perm) ;
+        } catch ( \Illuminate\Database\QueryException $e) {
+            return BF::result(false, $e->getMessage());
+        }
+
+        return BF::result(true, $result, '[usertype] index');
+
     }
 
     public function create()
@@ -20,7 +66,8 @@ class UserTypeController extends Controller
 
     public function store(Request $request)
     {
-        $input = BF::decodeInput($request->getContent());
+        $decode = BF::decodeInput($request->getContent());
+        $input = (array)$decode['data'] ;
         $rules = array(
             'name' => 'required|min:3|max:50',
         );
@@ -29,13 +76,18 @@ class UserTypeController extends Controller
             return BF::result(false, $validator->messages()->first());
         }
 
-        $testApi = true ;
-        if($testApi){
+        if(BF::TestSetting('userTypeAlert')){
+            //--- สำหรับเทส alert popup
+            return BF::result(false, "ชื่อซ้ำ: {$input['name']}");
+        }
+
+        if(BF::TestSetting('userTypeNotSave')){
             //--- สำหรับเทส ไม่ต้องคอยลบข้อมูลเวลาทดสอบฝั่ง client
         }else{
             try {
                 unset($input['token']);
                 $status = UserType::create($input);
+
                 if($status === NULL) {
                     return BF::result(false, 'failed!');
                 }
@@ -47,7 +99,7 @@ class UserTypeController extends Controller
             }
         }
         $data = [] ;
-        return BF::result(true, $data, 'usertype create');
+        return BF::result(true, $data, '[usertype] create');
     }
 
     public function show($id)
@@ -57,25 +109,36 @@ class UserTypeController extends Controller
     public function edit($id)
     {
         $data =  UserType::find($id);
-        return BF::result(true, ['action' => 'edit', 'data' => $data ]);
+        return BF::result(true, $data, '[usertype] edit');
     }
 
-    public function update($id)
+    public function update(Request $request,$id)
     {
         if(empty($id)) {
             return BF::result(false, 'ไม่พบข้อมูลนี้ค่ะ');
         }
-        $data = Input::all();
-        $data = array_diff_key($data, array_flip(['id','_method','deleted_at','deleted_by','updated_at','created_at']));
+        $decode = BF::decodeInput($request->getContent());
+        $input = (array)$decode['data'] ;
+        $input = array_diff_key($input, array_flip(['id','_method','deleted_at','deleted_by','updated_at','created_at']));
+        $rules = array(
+            'name' => 'required|min:3|max:50',
+        );
+        $validator = Validator::make($input, $rules);
+        if ($validator->fails()) {
+            return BF::result(false, $validator->messages()->first());
+        }
         //$data["updated_by"] = Session::get('user_id');
         try {
-            $status = UserType::whereId($id)->update($data);
+            $status = UserType::whereId($id)->update($input);
             if($status == 1) {
-                return BF::result(true, ['action' => 'update', 'id' => $id]);
+                $data = [
+                    'id' => $id
+                ] ;
+                return BF::result(true, $data, '[usertype] update');
             }
         } catch ( \Illuminate\Database\QueryException $e) {
             if($e->getCode() == 23000) {
-                return BF::result(false, "ชื่อซ้ำ: {$data['name']}");
+                return BF::result(false, "ชื่อซ้ำ: {$input['name']}");
             }
             return BF::result(false, $e->getMessage());
         }
@@ -84,16 +147,21 @@ class UserTypeController extends Controller
 
     public function destroy($id)
     {
-        if(empty($id)){
+        if(empty($id)) {
             return BF::result(false, 'ไม่พบข้อมูลนี้ค่ะ');
         }
         $data = UserType::find($id);
+        if(BF::TestSetting('userTypeNotDelete')){
+            //--- สำหรับเทส alert popup
+            return BF::result(true, $data, '[usertype] delete test');
+        }
+
         if (is_null($data)) {
             UserType::withTrashed()->whereId($id)->first()->restore();
-            return BF::result(true, ['action' => 'restore', 'id' => $id]);
+            return BF::result(true, $data, '[usertype] restore');
         }else{
             $data->delete();
-            return BF::result(true, ['action' => 'delete', 'id' => $id]);
+            return BF::result(true, $data, '[usertype] delete');
         }
     }
 }
